requirements.txt
Cap1xxx==0.1.3
ExplorerHAT==0.4.2
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.6.1
RPi.GPIO==0.6.3
RTIMULib==7.2.1
SimpleCV==1.3
Werkzeug==0.9.6
argparse==1.2.1
automationhat==0.0.4
blinker==1.3
blinkt==0.1.0
chardet==2.3.0
colorama==0.3.2
decorator==3.4.0
drumhat==0.0.5
envirophat==0.0.6
fourletterphat==0.0.2
gpiozero==1.4.0
html5lib==0.999
ipython==2.3.0
itsdangerous==0.24
lxkeymap==0.1
mcpi==0.1.1
microdotphat==0.1.3
mote==0.0.3
motephat==0.0.2
ndg-httpsclient==0.3.2
numpy==1.8.2
pexpect==3.2
phatbeat==0.1.0
pianohat==0.0.5
picamera==1.13
picraft==1.0
pifacecommon==4.2.1
pifacedigitalio==3.1.0
piglow==1.2.4
pigpio==1.35
pyOpenSSL==0.13.1
pyasn1==0.1.7
pygame==1.9.2a0
pygobject==3.14.0
pyinotify==0.9.4
pyparsing==2.2.0
pyserial==2.6
python-apt==0.9.3.12
rainbowhat==0.0.2
requests==2.4.3
scipy==0.14.0
scrollphat==0.0.7
scrollphathd==1.2.0
sense-emu==1.0
sense-hat==2.2.0
simplegeneric==0.8.1
six==1.8.0
skywriter==0.0.7
smbus==1.1
sn3218==1.2.7
spidev==3.0
svgwrite==1.1.12
touchphat==0.0.1
twython==3.1.2
urllib3==1.9.1
wheel==0.24.0
wsgiref==0.1.2
